pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"


[project]
name = "dex_api"
description = "Mercatus DEX REST API"
readme = "README.md"
dynamic = ["version"]
requires-python = ">=3.11"
keywords = ["pact", "kadena"]
license = {text = "BSD-3-Clause"}
dependencies = ["aiohttp>=3.10.11",
                "aioitertools>=0.11.0",
                "async-lru>=2.0.4",
                "flask-orjson>=2.0.0",
                "PyYAML>=6.0.0",
                "Quart>=0.19.9",
                "quart-cors>=0.7.0",
                "uvloop>=0.19.0",
                "motor>=3.5.0",
                "numpy>=2.1.0",
                "Hypercorn>=0.17.3"]

[tool.setuptools.dynamic]
version = {attr = "dex_api.version.VERSION"}

[tool.setuptools]
packages = ["dex_api"]
